Found bug in A* solver algorithm, even though the algorith I had technically worked it was not proper A*. I added a couple extra versions to try to fix it and make it work properly.

+++ MazeCell.cpp
@@ -25,12 +25,12 @@ std::size_t std::hash<CellCoords>::operator()(const CellCoords& coords) const {
 }
 
 // Returns the manhattan distance from cell_a to cell_b
-int CellCoords::manhattan_distance(const CellCoords& cell_a, const CellCoords& cell_b){
+double CellCoords::manhattan_distance(const CellCoords& cell_a, const CellCoords& cell_b){
     return std::abs(cell_a.row - cell_b.row) + std::abs(cell_a.col - cell_b.col);
 }
 
 // Returns the euclidean distance (stright line distance) from cell_a to cell_b
-int CellCoords::euclidean_distance(const CellCoords& cell_a, const CellCoords& cell_b){
+double CellCoords::euclidean_distance(const CellCoords& cell_a, const CellCoords& cell_b){
     return std::sqrt(std::pow(cell_a.row - cell_b.row, 2) + std::pow(cell_a.col - cell_b.col, 2));
 }
